Cleanup: keep managers/IdlePointerManager.swift; remove duplicates and update references
