+++ boringNotch/models/Constants.swift
@@ -81,6 +81,10 @@ extension Defaults.Keys {
     )
     static let nonNotchHeight = Key<CGFloat>("nonNotchHeight", default: 32)
     static let notchHeight = Key<CGFloat>("notchHeight", default: 32)
+    // Idle auto-collapse
+    static let idleEnabled = Key<Bool>("idleEnabled", default: true)
+    static let idleSeconds = Key<Double>("idleSeconds", default: 5.0)
+    static let idleTolerancePx = Key<Double>("idleTolerancePx", default: 2.0)
         //static let openLastTabByDefault = Key<Bool>("openLastTabByDefault", default: false)
     
         // MARK: Appearance
